Break up build_slides  a little more
